+++ lib/flutter_insider.dart
@@ -21,7 +21,7 @@ class FlutterInsider {
 
     args["appGroup"] = appGroup;
     args["partnerName"] = partnerName;
-    args["sdkVersion"] = "F-3.10.9+nh";
+    args["sdkVersion"] = "F-3.12.0+nh";
 
     if (customEndpoint != null) {
       args["customEndpoint"] = customEndpoint;